+++ index.ts
@@ -343,20 +343,20 @@ import { CircleBrush } from './src/brushes/circle_brush.class';
 import { SprayBrush } from './src/brushes/spray_brush.class';
 import { PatternBrush } from './src/brushes/pattern_brush.class';
 import { FabricObject as Object } from './src/shapes/Object/FabricObject';
-import { Line } from './src/shapes/line.class';
-import { Circle } from './src/shapes/circle.class';
-import { Triangle } from './src/shapes/triangle.class';
-import { Ellipse } from './src/shapes/ellipse.class';
-import { Rect } from './src/shapes/rect.class';
-import { Path } from './src/shapes/path.class';
-import { Polyline } from './src/shapes/polyline.class';
-import { Polygon } from './src/shapes/polygon.class';
+import { Line } from './src/shapes/Line';
+import { Circle } from './src/shapes/Circle';
+import { Triangle } from './src/shapes/Triangle';
+import { Ellipse } from './src/shapes/Ellipse';
+import { Rect } from './src/shapes/Rect';
+import { Path } from './src/shapes/Path';
+import { Polyline } from './src/shapes/Polyline';
+import { Polygon } from './src/shapes/Polygon';
 import { Text } from './src/shapes/Text/Text';
 import { IText } from './src/shapes/IText/IText';
-import { Textbox } from './src/shapes/textbox.class';
-import { Group } from './src/shapes/group.class';
-import { ActiveSelection } from './src/shapes/active_selection.class';
-import { Image } from './src/shapes/image.class';
+import { Textbox } from './src/shapes/Textbox';
+import { Group } from './src/shapes/Group';
+import { ActiveSelection } from './src/shapes/ActiveSelection';
+import { Image } from './src/shapes/Image';
 import { getEnv, getDocument, getWindow, setEnvForTests } from './src/env';
 import { createCollectionMixin } from './src/Collection';
 import { parseAttributes } from './src/parser/parseAttributes';
