+++ src/controls/default_controls.ts
@@ -1,5 +1,5 @@
 import { FabricObject } from '../shapes/Object/FabricObject';
-import { Textbox } from '../shapes/textbox.class';
+import { Textbox } from '../shapes/Textbox';
 import { scaleCursorStyleHandler, scalingEqually } from './scale';
 import { changeWidth } from './changeWidth';
 import { rotationStyleHandler, rotationWithSnapping } from './rotate';
