+++ src/controls/polyControl.ts
@@ -2,7 +2,7 @@ import { Point } from '../point.class';
 import { Control } from './control.class';
 import { TMat2D } from '../typedefs';
 import { iMatrix } from '../constants';
-import type { Polyline } from '../shapes/polyline.class';
+import type { Polyline } from '../shapes/Polyline';
 import { multiplyTransformMatrices } from '../util/misc/matrix';
 import {
   TPointerEvent,
